:recycle: Refactor landing page by creating modular components for header, hero, government services, statistics, contact, and footer. This improves code organization and maintainability.
